Reorganized project folders and files
