+++ config/webpack.config.ts
@@ -2,14 +2,12 @@ import * as fs from "fs";
 import * as path from "path";
 import * as webpack from "webpack";
 import * as CleanPlugin from "clean-webpack-plugin";
-import * as CopyPlugin from "copy-webpack-plugin";
 import * as NodemonPlugin from "nodemon-webpack-plugin";
 import * as PermissionsPlugin from "webpack-permissions-plugin";
 // tslint:disable-next-line
 const tsconfig = require("../tsconfig.json");
 const root = path.resolve(__dirname, "..");
 const paths = {
-    assets: "public",
     dist: `${root}/${tsconfig.compilerOptions.outDir}`,
     entry: `${root}/src/bin/www.ts`,
     license: "copyright-notice.txt",
@@ -62,9 +60,6 @@ export default {
                 verbose: false
             }
         ),
-        new CopyPlugin([
-            { from: `${paths.source}/${paths.assets}`, to: `${paths.dist}/${paths.assets}` }
-        ]),
         new NodemonPlugin(),
         new webpack.BannerPlugin(
             { banner: fs.readFileSync(paths.license).toString() }
