+++ package.json
@@ -49,7 +49,6 @@
     "js-yaml": "3.10.0",
     "moment": "2.20.1",
     "morgan": "1.9.0",
-    "serve-favicon": "2.4.5",
     "uuid": "3.2.1",
     "winston": "2.4.0",
     "xml2js": "0.4.19"
@@ -60,26 +59,26 @@
     "@types/eureka-js-client": "4.3.1",
     "@types/express": "4.11.1",
     "@types/helmet": "0.0.37",
-    "@types/jest": "22.1.2",
+    "@types/jest": "22.1.3",
     "@types/js-yaml": "3.10.1",
     "@types/morgan": "1.7.35",
     "@types/node": "9.4.6",
-    "@types/serve-favicon": "2.2.30",
     "@types/sinon": "4.1.3",
     "@types/supertest": "2.0.4",
     "@types/uuid": "3.4.3",
+    "@types/webpack": "3.8.8",
     "@types/winston": "2.3.8",
     "@types/xml2js": "0.4.2",
     "ajv": "6.1.1",
     "clean-webpack-plugin": "0.1.18",
     "copy-webpack-plugin": "4.4.1",
-    "jest": "22.3.0",
+    "jest": "22.4.0",
     "nodemon-webpack-plugin": "0.1.6",
     "sinon": "4.3.0",
     "supertest": "3.0.0",
     "ts-jest": "22.0.4",
     "ts-loader": "3.5.0",
-    "ts-node": "4.1.0",
+    "ts-node": "5.0.0",
     "tslint": "5.9.1",
     "typescript": "2.7.2",
     "webpack": "3.11.0",
@@ -92,7 +91,6 @@
     "access": "public"
   },
   "jest": {
-    "mapCoverage": true,
     "moduleFileExtensions": [
       "js",
       "json",
