+++ src/parsers/index.ts
@@ -1,0 +1,12 @@
+import { GenericParser } from "./generic.parser";
+import { JSONParser } from "./json";
+import { PlainTextParser } from "./plain-text";
+import { XMLParser } from "./xml";
+import { YAMLParser } from "./yaml";
+export {
+    GenericParser,
+    JSONParser,
+    PlainTextParser,
+    XMLParser,
+    YAMLParser
+};
