+++ src/routers/index.ts
@@ -1,0 +1,8 @@
+import { GenericRouter } from "./generic.router";
+import { Router as EurekaClientRouter } from "./eureka-client";
+import { Router as ConfigReaderRouter } from "./config-reader";
+export {
+    GenericRouter,
+    EurekaClientRouter,
+    ConfigReaderRouter
+};
