+++ src/services/index.ts
@@ -1,0 +1,8 @@
+import { Eureka } from "./eureka-client";
+import { FileReaderService } from "./file-reader";
+import * as logger from "./logger";
+export {
+    Eureka,
+    FileReaderService,
+    logger
+};
