+++ src/utils/index.ts
@@ -1,0 +1,6 @@
+import * as AppUtil from "./app";
+import * as ServerUtil from "./server";
+export {
+    AppUtil,
+    ServerUtil
+};
