+++ tsconfig.json
@@ -4,9 +4,6 @@
     "module": "commonjs",
     "moduleResolution": "node",
     "outDir": "dist",
-    "sourceMap": true,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
     "noUnusedLocals": true
   },
   "include": [
